+++ frontend/src/api/post.api.ts
@@ -17,16 +17,16 @@ export const postAPI = {
         data.append('title', dto.title)
         data.append('content', dto.content)
         data.append('userId', String(dto.userId))
-        data.append('image', dto.image as string)
-        return instance.post<IApiOkResponse<PostDto[]>>('/posts', dto).then((response) => response.data)
+        data.append('image', dto.image[0])
+        return instance.post<IApiOkResponse<PostDto[]>>('/posts', data).then((response) => response.data)
     },
     updatePost(dto: UpdatePostDto): Promise<IApiOkResponse<PostDto[]>> {
         const data = new FormData();
         data.append('title', dto.title)
         data.append('content', dto.content)
         data.append('likesCount', String(dto.likesCount))
-        data.append('image', dto.image)
-        return instance.patch<IApiOkResponse<PostDto[]>>(`/posts${dto.id}`, dto).then((response) => response.data)
+        data.append('image', dto.image[0])
+        return instance.patch<IApiOkResponse<PostDto[]>>(`/posts${dto.id}`, data).then((response) => response.data)
     },
     deletePost(id: number): Promise<IApiOkResponse<PostDto[]>> {
         return instance.delete<IApiOkResponse<PostDto[]>>(`posts/${id}`).then((response) => response.data)