+++ frontend/src/types/post/updatePost.dto.ts
@@ -2,6 +2,6 @@ export interface UpdatePostDto {
     id: number
     title: string;
     content: string;
-    image: string;
+    image: any;
     likesCount: number;
 }